requirements.txt
boto3
django-celery-beat==2.4.0
django==4.0.7
celery==5.2.7
redis==4.3.4
flower==1.2.0
psycopg2-binary==2.9.2    
django-enumfield==3.1
Pillow
django-jazzmin
djangorestframework
drf-spectacular[sidecar]
mercadopago
pytest-django
pytest-cov
pytest-mock